feat: implement calculator logic, button structure
Add buttons in a grid layout. The backspace and clear buttons can be
used to further manipulate the numbers in the display.

This approach utilises a finite state machine, where the 5 types of
input (clear, backspace, digit, operator, equals) are used to
transition between 4 states (FIRST, OPERATOR, SECOND, RESULT).

The current stage is shown for debug purposes.